Lots less warnings. More typecasting & fixs

+++ Kernel/hdd/diskio.c
@@ -39,7 +39,7 @@ uint8 device_read(uint16* dest, uint32 offset_sectors, uint32 num_sectors, disk_
 		  num_sectors = num_sectors - 256;
 		  if (pio_28bit_read(device, dest, offset_sectors, 0) == 0) return 0; //Error, fail to read
 		  offset_sectors = offset_sectors + 256; //Increment the offset by 256
-		  dest = ((uint32*)dest) + (sizeof(uint16) * 256); //Increment the destination pointer by that
+		  dest = ((uint16*)dest) + (sizeof(uint16) * 256); //Increment the destination pointer by that
 		} else if (num_sectors < 1)
 		{
 		  break;
